move more of bootstrapping into rootblock
