+++ model/bootstrap/filenames.go
@@ -25,6 +25,8 @@ var (
 
 	PathRootBlockData             = filepath.Join(DirnamePublicBootstrap, "root-block.json")
 	PathRootEpoch                 = filepath.Join(DirnamePublicBootstrap, "root-epoch.json")
+	PathRootResult                = filepath.Join(DirnamePublicBootstrap, "root-result.json")
+	PathRootSeal                  = filepath.Join(DirnamePublicBootstrap, "root-seal.json")
 	PathRootProtocolStateSnapshot = filepath.Join(DirnamePublicBootstrap, "root-protocol-state-snapshot.json")
 
 	FilenameWALRootCheckpoint = "root.checkpoint"
